Move testDefaultRequestLimit over to retry file as it uses the same project
