+++ google-cloud-bigquerystorage/src/test/java/com/google/cloud/bigquery/storage/v1/it/ITBigQueryWriteManualClientTest.java
@@ -1608,74 +1608,4 @@ public class ITBigQueryWriteManualClientTest {
       assertEquals("50", queryIter.next().get(0).getStringValue());
     }
   }
-
-  @Test
-  public void testDefaultRequestLimit()
-      throws IOException, InterruptedException, ExecutionException {
-    DatasetId datasetId =
-        DatasetId.of("bq-write-api-java-retry-test", RemoteBigQueryHelper.generateDatasetName());
-    DatasetInfo datasetInfo = DatasetInfo.newBuilder(datasetId).build();
-    bigquery.create(datasetInfo);
-    try {
-      String tableName = "no_error_table";
-      TableId tableId = TableId.of(datasetId.getProject(), datasetId.getDataset(), tableName);
-      Field col1 = Field.newBuilder("col1", StandardSQLTypeName.STRING).build();
-      Schema originalSchema = Schema.of(col1);
-      TableInfo tableInfo =
-          TableInfo.newBuilder(tableId, StandardTableDefinition.of(originalSchema)).build();
-      bigquery.create(tableInfo);
-      TableName parent = TableName.of(datasetId.getProject(), datasetId.getDataset(), tableName);
-      try (StreamWriter streamWriter =
-          StreamWriter.newBuilder(parent.toString() + "/_default")
-              .setWriterSchema(CreateProtoSchemaWithColField())
-              .setRetrySettings(null)
-              .build()) {
-        ApiFuture<AppendRowsResponse> response =
-            streamWriter.append(
-                CreateProtoRows(
-                    new String[] {new String(new char[19 * 1024 * 1024]).replace("\0", "a")}));
-        try {
-          AppendRowsResponse resp = response.get();
-          LOG.info(
-              "Message succeded.  Dataset info: "
-                  + datasetInfo.toString()
-                  + " tableinfo: "
-                  + tableInfo.toString()
-                  + " parent: "
-                  + parent
-                  + "streamWriter: "
-                  + streamWriter.toString()
-                  + "response: "
-                  + resp);
-          Assert.fail("Large request should fail with InvalidArgumentError");
-        } catch (ExecutionException ex) {
-          LOG.info(
-              "Message failed.  Dataset info: "
-                  + datasetInfo.toString()
-                  + " tableinfo: "
-                  + tableInfo.toString()
-                  + " parent: "
-                  + parent
-                  + "streamWriter: "
-                  + streamWriter);
-          assertEquals(io.grpc.StatusRuntimeException.class, ex.getCause().getClass());
-          io.grpc.StatusRuntimeException actualError =
-              (io.grpc.StatusRuntimeException) ex.getCause();
-          // This verifies that the Beam connector can consume this custom exception's grpc
-          // StatusCode
-          // TODO(yiru): temp fix to unblock test, while final fix is being rolled out.
-          if (actualError.getStatus().getCode() != Code.INTERNAL) {
-            assertEquals(Code.INVALID_ARGUMENT, actualError.getStatus().getCode());
-            assertThat(
-                actualError
-                    .getStatus()
-                    .getDescription()
-                    .contains("AppendRows request too large: 19923131 limit 10485760"));
-          }
-        }
-      }
-    } finally {
-      RemoteBigQueryHelper.forceDelete(bigquery, datasetId.toString());
-    }
-  }
 }
